reorganize pages and components
