+++ frontend/src/components/Footer.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 .site-footer {
   width: 100%;
