+++ frontend/src/components/NavBar.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 .navbar {
   background: var(--theme-black);
