+++ frontend/src/pages/Blog.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Blog Page Basics */
