+++ frontend/src/pages/BlogDetail.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Blog Detail Page Basics */
