+++ frontend/src/pages/BlogDetail.js
