+++ frontend/src/pages/Contact.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Contact Page Basics */
