+++ frontend/src/pages/Contact.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
-import './App.css';
-import emailLogo from './assets/socials/email-logo.png';
-import linkedinLogo from './assets/socials/linkedin-logo.png';
-import gitLogo from './assets/socials/git-logo.png';
+import '../App.css';
+import emailLogo from '../assets/socials/email-logo.png';
+import linkedinLogo from '../assets/socials/linkedin-logo.png';
+import gitLogo from '../assets/socials/git-logo.png';
 import './Contact.css';
 import Lottie from 'lottie-react';
-import mapPointer from './assets/animations/map_pointer.json';
+import mapPointer from '../assets/animations/map_pointer.json';
 
 function Contact() {
     const [message, setMessage] = useState("");
