+++ frontend/src/pages/Home.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Home Page Basics */
