+++ frontend/src/pages/Home.js
@@ -1,31 +1,31 @@
 import React, { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import Lottie from 'lottie-react';
-import foxHello from './assets/animations/fox_hello.json';
+import foxHello from '../assets/animations/fox_hello.json';
 import './Home.css';
-import './NavBar.css';
-import reactIcon from './assets/skills/react.svg';
-import htmlIcon from './assets/skills/html.svg';
-import javascriptIcon from './assets/skills/javascript.svg';
-import blazorIcon from './assets/skills/blazor.svg';
-import bootstrapIcon from './assets/skills/bootstrap.svg';
-import dotnetIcon from './assets/skills/dotnet.svg';
-import databaseIcon from './assets/skills/database.svg';
-import springbootIcon from './assets/skills/springboot.svg';
-import awsIcon from './assets/skills/aws.svg';
-import dockerIcon from './assets/skills/docker.svg';
-import jenkinsIcon from './assets/skills/jenkins.svg';
-import agileIcon from './assets/skills/agile.svg';
-import oopIcon from './assets/skills/OOP.svg';
-import tddIcon from './assets/skills/TDD.svg';
-import vsIcon from './assets/skills/VS.svg';
-import intellijIcon from './assets/skills/intellijidea.svg';
-import githubIcon from './assets/skills/github.svg';
-import atlassianIcon from './assets/skills/atlassian.svg';
-import profilePhoto from './assets/global/profile.jpeg';
-import opusLogo from './assets/history/opus-logo.png';
-import cigLogo from './assets/history/cig-logo.jpeg';
-import ewuLogo from './assets/history/ewu-logo.svg';
+import '../components/NavBar.css';
+import reactIcon from '../assets/skills/react.svg';
+import htmlIcon from '../assets/skills/html.svg';
+import javascriptIcon from '../assets/skills/javascript.svg';
+import blazorIcon from '../assets/skills/blazor.svg';
+import bootstrapIcon from '../assets/skills/bootstrap.svg';
+import dotnetIcon from '../assets/skills/dotnet.svg';
+import databaseIcon from '../assets/skills/database.svg';
+import springbootIcon from '../assets/skills/springboot.svg';
+import awsIcon from '../assets/skills/aws.svg';
+import dockerIcon from '../assets/skills/docker.svg';
+import jenkinsIcon from '../assets/skills/jenkins.svg';
+import agileIcon from '../assets/skills/agile.svg';
+import oopIcon from '../assets/skills/OOP.svg';
+import tddIcon from '../assets/skills/TDD.svg';
+import vsIcon from '../assets/skills/VS.svg';
+import intellijIcon from '../assets/skills/intellijidea.svg';
+import githubIcon from '../assets/skills/github.svg';
+import atlassianIcon from '../assets/skills/atlassian.svg';
+import profilePhoto from '../assets/global/profile.jpeg';
+import opusLogo from '../assets/history/opus-logo.png';
+import cigLogo from '../assets/history/cig-logo.jpeg';
+import ewuLogo from '../assets/history/ewu-logo.svg';
 
 function Home() {
