+++ frontend/src/pages/NotFound.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Not Found Page Styles */
