+++ frontend/src/pages/ProjectDetail.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Project Detail Page Basics */
