+++ frontend/src/pages/ProjectDetail.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import projects from "./data/projects.json";
-import portfolioImg from './assets/projects/portfolio.png';
-import comingSoonImg from './assets/projects/comingsoon.jpeg';
+import projects from "../data/projects.json";
+import portfolioImg from '../assets/projects/portfolio.png';
+import comingSoonImg from '../assets/projects/comingsoon.jpeg';
 import "./ProjectDetail.css";
-import linkLogoWhite from './assets/links/link-logo-white.png';
-import gitLogoWhite from './assets/links/git-logo-white.png';
+import linkLogoWhite from '../assets/links/link-logo-white.png';
+import gitLogoWhite from '../assets/links/git-logo-white.png';
 
 function ProjectDetail() {
   const { id } = useParams();
