+++ frontend/src/pages/Projects.css
@@ -1,4 +1,4 @@
-@import './styles/theme.css';
+@import '../styles/theme.css';
 
 /* Projects Page Basics */
