+++ frontend/src/pages/Projects.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import projects from "./data/projects.json";
+import projects from "../data/projects.json";
 import "./Projects.css";
-import portfolioImg from './assets/projects/portfolio.png';
-import comingSoonImg from './assets/projects/comingsoon.jpeg';
+import portfolioImg from '../assets/projects/portfolio.png';
+import comingSoonImg from '../assets/projects/comingsoon.jpeg';
 
 function Projects() {
   // Map project IDs to imported images
